Split columns UI code into files
