+++ DisguiseUnityRenderStream/Editor/Parameters/TreeView/Columns/ComponentColumn.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5851ea6de9a64f04bbe638bcb96873e3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
