+++ DisguiseUnityRenderStream/Editor/Parameters/TreeView/Columns/NameColumn.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 810f79a53e7be31468c343acc3d1b968
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
