+++ DisguiseUnityRenderStream/Editor/Parameters/TreeView/Columns/ObjectColumn.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a896ced835bd5334aa1c49d1c328c97a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
